Патч от 02.02.2023. Client.
feature: реализована полноценная функциональность работы с корзиной пользователя (пока не реализована авторизация id корзины берётся из переменных окружения).

feature: полностью переработаны классы для продуктов, создан базовый класс Product и наследующие от него ShowcaseProduct и CartProduct.

feature: добавлена временная переменная окружения для работы с корзиной пользователя - CART_ID.

feature: полностью реализована витрина продуктов для главной страницы и страницы категории продуктов.

feature: ссылки на картинки для различных компонентов вынесены в отдельный файл с константами.

feature: удалены константы параметров заказа.

feature: полностью переработаны хранилища корзины и данных о доставке и заказе.

feature: UrlQueryHandler -> ParsedUrlQueryHandler.

feature: стилизованные компоненты в styles вынесены в отдельную папку внутри styles - styled-components.

feature: сервисы в отдельных модулях вынесены в папки [moduleName]/api.

feature: для основного layout`a приложения создан базовый api-запрос, который запрашивает определенный пул данных. Используется для страниц с основным layout`ом.

feature: компоненты кнопок, которые имели логику, были вынесены из ui.

feature: добавлен ui-компонент carousel.

feature: добавлен новый хук - useScrollPosition.

feature: компонент OrderButton вынесен в отдельную папку.

feature: новые параметры для NumberFormatter.

fix: исправлены имена файлов types.

fix: в стилях для CartPanelItem теперь определяется последний ли элемент в списке с помощью CSS.

+++ client/src/api/models/pagination-params.model.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod'
 
 export const PaginationParamsSchema = z.object({
-  skip: z.number().min(0).optional().default(0),
-  take: z.number().min(0).optional().default(0)
+  skip: z.number().int().min(0).optional().default(0),
+  take: z.number().int().min(0).optional().default(0)
 })
 
 export type PaginationParams = z.infer<typeof PaginationParamsSchema>
